+++ lib_acl_cpp/include/acl_cpp/memcache/memcache_pool.hpp
@@ -19,10 +19,10 @@ public:
 	/**
 	 * 构造函数
 	 * @param addr {const char*} 服务端地址，格式：ip:port
-	 * @param count {int} 连接池的最大连接数限制
+	 * @param count {size_t} 连接池的最大连接数限制
 	 * @param idx {size_t} 该连接池对象在集合中的下标位置(从 0 开始)
 	 */
-	memcache_pool(const char* addr, int count, size_t idx = 0);
+	memcache_pool(const char* addr, size_t count, size_t idx = 0);
 	~memcache_pool();
 
 	/**
