+++ lib_acl_cpp/samples/connect_manager/connect_manager.cpp
@@ -11,7 +11,7 @@ connect_manager::~connect_manager()
 }
 
 acl::connect_pool* connect_manager::create_pool(const char* addr,
-	int count, size_t idx)
+	size_t count, size_t idx)
 {
 	return new connect_pool(addr, count, idx);
 }
