+++ lib_acl_cpp/src/http/http_request_pool.cpp
@@ -6,7 +6,7 @@ namespace acl
 {
 
 http_request_pool::http_request_pool(const char* addr,
-	int count, size_t idx /* = 0 */)
+	size_t count, size_t idx /* = 0 */)
 : connect_pool(addr, count, idx)
 {
 	conn_timeout_ = 30;
